style: renamed files and edited comments
