+++ client/app.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import {Navbar} from './components'
+import Navbar from './components/Navbar'
 import Routes from './routes'
 
 const App = () => {
