+++ client/components/AllProperties.js
