+++ client/components/Index.js
@@ -1,0 +1,9 @@
+export {default as Navbar} from './Navbar'
+export {default as UserHome} from './UserHome'
+export {Login, Signup} from './AuthForm'
+export {default as AllProperties} from './AllProperties'
+export {default as SingleProperty} from './SingleProperty'
+export {default as AllUsers} from './AllUsers'
+export {default as PastOrders} from './PastOrders'
+export {default as Cart} from './Cart'
+export {default as LandingPage} from './LandingPage'
