+++ client/components/UserHome.js
@@ -1,0 +1,21 @@
+import React from 'react'
+// import PropTypes from 'prop-types' //SAFEWORD: PICKLES
+import {connect} from 'react-redux'
+
+export const UserHome = props => {
+  const {email} = props
+
+  return (
+    <div>
+      <h3>Welcome, {email}</h3>
+    </div>
+  )
+}
+
+const mapState = state => {
+  return {
+    email: state.user.email
+  }
+}
+
+export default connect(mapState)(UserHome)
