--- client/components/index.js
@@ -1,14 +1,0 @@
-/**
- * `components/index.js` exists simply as a 'central export' for our components.
- * This way, we can import all of our components from the same place, rather than
- * having to figure out which file they belong to!
- */
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
-export {default as AllProperties} from './all-properties'
-export {default as SinglePropery} from './single-property'
-export {default as AllUsers} from './all-users'
-export {default as PastOrders} from './past-orders'
-export {default as Cart} from './cart'
-export {default as LandingPage} from './landing-page'
